Popup positioning: wip
